Misc refactor and fitness tweaks.

+++ summarize_results.py
@@ -8,7 +8,7 @@ import seaborn as sns
 
 import experiments
 import gif_files
-import kernel
+from kernel import simulate_organism, FitnessGoal, Simulator
 
 
 SPECIES_INDEX_COLUMNS = (
@@ -30,12 +30,12 @@ ORGANISM_INDEX_COLUMNS = (
 
 def get_random_population(species_list):
     num_species = len(species_list)
-    simulator = kernel.Simulator(num_species, 1, experiments.NUM_ORGANISMS)
+    simulator = Simulator(num_species, 1, experiments.NUM_ORGANISMS)
     simulator.populate(np.array([
         species_data.phenotype_program.serialize()
         for species_data in species_list]))
-    # TODO: NONE fitness goal? Fitness doesn't matter.
-    videos = simulator.simulate_and_record(kernel.FitnessGoal.STILL_LIFE)
+    # TODO: Maybe use a no-op fitness goal, since it doesn't matter here?
+    videos = simulator.simulate_and_record(FitnessGoal.STILL_LIFE)
     # Get rid of the pointless trial index, since we only ran one trial.
     return videos[:, 0]
 
@@ -60,23 +60,17 @@ def summarize_species_data(species_data, species_path):
     with open(species_path.joinpath('phenotype_program.txt'), 'w') as file:
         file.write(str(species_data.phenotype_program))
 
-    # Output the samples of the best organisms at multiple points in species
-    # evolution across all trials.
-    for generation, samples_by_trial in species_data.sample_organisms.items():
-        sample_path = species_path.joinpath(
-            f'best_organisms_from_gen{generation:03d}')
-        sample_path.mkdir()
-        for trial, organism_data in enumerate(samples_by_trial):
-            np.save(
-                sample_path.joinpath(f'genotype_from_trial{trial:d}.npy'),
-                organism_data.genotype)
-            gif_files.save_image(
-                kernel.simulate_organism(
-                    species_data.phenotype_program.serialize(),
-                    organism_data.genotype),
-                sample_path.joinpath(
-                    f'simulation_from_trial{trial:d}'
-                    f'_f{organism_data.fitness_scores[-1]}.gif'))
+    # Save the genotype and video for the best organism from this species.
+    best_organism = species_data.best_organism
+    np.save(
+        species_path.joinpath('best_organism_genotype.npy'),
+        best_organism.genotype)
+    gif_files.save_image(
+        simulate_organism(
+            species_data.phenotype_program.serialize(),
+            best_organism.genotype),
+        species_path.joinpath(
+            f'best_organism_f{best_organism.fitness_scores[-1]}.gif'))
 
 
 def recursive_delete_directory(path):
@@ -88,15 +82,17 @@ def recursive_delete_directory(path):
     path.rmdir()
 
 
-def summarize_experiment_data(experiment_data, experiment_path):
+def summarize_experiment_data(experiment):
     # Go through all files in the output directory for this experiment and all
     # of its subdirectories and delete any generated output (that is,
     # everything but the raw experiment data in state.pickle)
-    for file in experiment_path.glob('*'):
+    for file in experiment.path.glob('*'):
         if file.is_dir():
             recursive_delete_directory(file)
-        elif file.name != 'state.pickle':
+        elif file.name != 'state.pickle' and file.name != 'results.pickle':
             file.unlink()
+
+    experiment_data = experiment.get_data()
 
     # Generate a chart summarizing species fitness across trials
     species_fitness_by_trial = pd.DataFrame(
@@ -108,25 +104,50 @@ def summarize_experiment_data(experiment_data, experiment_path):
         data=species_fitness_by_trial, kind='line',
         x='Generation', y='Fitness', hue='Trial')
     fig.set(title='Species Fitness by Trial')
-    fig.savefig(experiment_path.joinpath('species_fitness.svg'))
+    fig.savefig(experiment.path.joinpath('species_fitness.svg'))
     plt.close()
 
     # Summarize experiment results for the best species evolved in each trial,
     # including a sample initial population.
     random_populations = get_random_population(
         experiment_data.best_species_per_trial)
-    for trial in range(experiment_data.trial + 1):
+    best_fitness = 0
+    best_video = None
+    for trial in range(experiment.trial + 1):
+        # Generate output for this species
         species_data = experiment_data.best_species_per_trial[trial]
-        species_path = experiment_path.joinpath(
+        species_path = experiment.path.joinpath(
             f'best_species_from_trial_{trial:d}'
             f'_f{species_data.fitness_scores[-1]}')
         species_path.mkdir()
         summarize_species_data(species_data, species_path)
+
+        # Save the best video for each trial, and track best overall
+        best_organism = species_data.best_organism
+        video = simulate_organism(
+            species_data.phenotype_program.serialize(),
+            best_organism.genotype)
+        fitness = best_organism.fitness_scores[-1]
+        gif_files.save_image(
+            video,
+            experiment.path.joinpath(
+                f'best_organism_from_trial_{trial}_f{fitness}.gif'))
+        if fitness > best_fitness:
+            best_fitness = fitness
+            best_video = video
+
+        # Save a sample random population to visualize this species.
         population_path = species_path.joinpath('random_initial_population')
         population_path.mkdir()
         for index, video in enumerate(random_populations[trial]):
             gif_files.save_image(
                 video, population_path.joinpath(f'sample_{index:02d}.gif'))
+
+    # Save the overall best organism found for this experiment.
+    gif_files.save_image(
+        best_video,
+        experiment.path.parent.joinpath(
+            f'{experiment.name}_f{best_fitness}.gif'))
 
 
 def main():
@@ -148,9 +169,8 @@ def main():
         # If the summary data is out of date or the user requested a full
         # rebuild from the command line.
         if data_update_time > summary_update_time or args.rebuild:
-            experiment_path = experiment.state_path.parent
-            experiment_path.mkdir(exist_ok=True)
-            summarize_experiment_data(experiment, experiment_path)
+            experiment.path.mkdir(exist_ok=True)
+            summarize_experiment_data(experiment)
     # TODO: If all experiments have at least one trial done, generate charts
     # comparing performance between experiments.
